workflow fixes: added missing fields, $schema prop, validate flag + moved validator to core, fetch schema from permanent url
