+++ javascript/core/docs/core.api.md
@@ -781,6 +781,8 @@ export interface SpreadArgs extends InputColumnListArgs {
 // @public (undocumented)
 export interface Step<T extends object | void | unknown = unknown> {
     args: T;
+    // (undocumented)
+    description?: string;
     id: string;
     input: {
         others?: NamedPortBinding[];
@@ -793,6 +795,8 @@ export interface Step<T extends object | void | unknown = unknown> {
 // @public (undocumented)
 export interface StepInput<T extends object | void | unknown = unknown> {
     args?: T;
+    // (undocumented)
+    description?: string;
     id?: string;
     input?: string | ({
         others?: PortBinding[];
@@ -993,7 +997,7 @@ export enum WindowFunction {
 //
 // @public
 export class Workflow {
-    constructor(workflowJson?: WorkflowObject);
+    constructor(workflowJson?: WorkflowObject, validate?: boolean);
     // (undocumented)
     addInput(input: string): void;
     // (undocumented)
@@ -1035,6 +1039,10 @@ export class Workflow {
 // @public (undocumented)
 export interface WorkflowObject {
     // (undocumented)
+    $id?: string;
+    // (undocumented)
+    $schema?: string;
+    // (undocumented)
     description?: string;
     // (undocumented)
     input?: string[];
@@ -1045,6 +1053,19 @@ export interface WorkflowObject {
     // (undocumented)
     steps: StepInput[];
 }
+
+// Warning: (ae-missing-release-tag) "WorkflowSchema" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export class WorkflowSchema {
+    // (undocumented)
+    isValid(workflowJson?: WorkflowObject): Promise<boolean>;
+}
+
+// Warning: (ae-missing-release-tag) "WorkflowSchemaInstance" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export const WorkflowSchemaInstance: WorkflowSchema;
 
 // (No @packageDocumentation comment for this package)
