+++ javascript/core/src/engine/index.ts
@@ -4,4 +4,5 @@
  */
 export * from './graph.js'
 export * from './GraphManager.js'
+export * from './validator.js'
 export * from './Workflow.js'
