+++ javascript/react/package.json
@@ -37,7 +37,6 @@
 	},
 	"dependencies": {
 		"@data-wrangling-components/core": "workspace:^",
-		"@data-wrangling-components/schema": "workspace:^",
 		"@data-wrangling-components/utilities": "workspace:^",
 		"@data-wrangling-components/verb-guidance": "workspace:^",
 		"@essex/arquero": "^3.0.0",
