+++ scripts/copy-schemas.mjs
@@ -9,7 +9,9 @@ import { fileURLToPath } from 'url'
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = dirname(__filename)
 
-const schemaReleases = path.join(__dirname, '../schema/releases')
+const latestWorkflowSchema = 'workflow.json'
+const schemaDir = path.join(__dirname, '../schema')
+const schemaReleases = path.join(schemaDir, 'releases')
 const webappVersionDirectory = path.join(
 	__dirname,
 	'../javascript/webapp/public/schema',
@@ -18,6 +20,11 @@ const webappVersionDirectory = path.join(
 if (!fs.existsSync(webappVersionDirectory)) {
 	fs.mkdirSync(webappVersionDirectory)
 }
+
+fs.copyFileSync(
+	path.join(schemaDir, latestWorkflowSchema),
+	path.join(webappVersionDirectory, latestWorkflowSchema),
+)
 
 fs.readdirSync(schemaReleases).forEach(file => {
 	const renamedFile = file.replace(/workflow-/, '')
